requirements.txt
Flask
gunicorn
flask_cors
werkzeug
Pillow==7.2.0
fastai==1.6.0